fix: refactor NewConfig, NewAppConstructor and remove duplicated s.SetupTest
