+++ x/farming/client/testutil/cli_helpers.go
@@ -3,48 +3,15 @@ package testutil
 import (
 	"fmt"
 
-	dbm "github.com/tendermint/tm-db"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	servertypes "github.com/cosmos/cosmos-sdk/server/types"
-	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/simapp/params"
-	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	"github.com/cosmos/cosmos-sdk/testutil"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
-	"github.com/cosmos/cosmos-sdk/testutil/network"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	bankcli "github.com/cosmos/cosmos-sdk/x/bank/client/cli"
 
-	crescentapp "github.com/crescent-network/crescent/app"
 	crescentcli "github.com/crescent-network/crescent/x/farming/client/cli"
 )
-
-// NewConfig returns config that defines the necessary testing requirements
-// used to bootstrap and start an in-process local testing network.
-func NewConfig(dbm *dbm.MemDB) network.Config {
-	encCfg := simapp.MakeTestEncodingConfig()
-
-	cfg := network.DefaultConfig()
-	cfg.AppConstructor = NewAppConstructor(encCfg, dbm)                   // the ABCI application constructor
-	cfg.GenesisState = crescentapp.ModuleBasics.DefaultGenesis(cfg.Codec) // farming genesis state to provide
-	return cfg
-}
-
-// NewAppConstructor returns a new network AppConstructor.
-func NewAppConstructor(encodingCfg params.EncodingConfig, db *dbm.MemDB) network.AppConstructor {
-	return func(val network.Validator) servertypes.Application {
-		return crescentapp.NewCrescentApp(
-			val.Ctx.Logger, db, nil, true, make(map[int64]bool), val.Ctx.Config.RootDir, 0,
-			crescentapp.MakeEncodingConfig(),
-			simapp.EmptyAppOptions{},
-			baseapp.SetPruning(storetypes.NewPruningOptionsFromString(val.AppConfig.Pruning)),
-			baseapp.SetMinGasPrices(val.AppConfig.MinGasPrices),
-		)
-	}
-}
 
 var commonArgs = []string{
 	fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
