+++ x/farming/client/testutil/cli_test.go
@@ -9,6 +9,10 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
+func TestIntegrationTestSuite(t *testing.T) {
+	suite.Run(t, new(IntegrationTestSuite))
+}
+
 func TestQueryCmdTestSuite(t *testing.T) {
 	suite.Run(t, new(QueryCmdTestSuite))
 }
